requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
watchdog>=3.0.0
python-dotenv>=1.0.0
GitPython>=3.1.40
numpy>=1.24.0
scikit-learn>=1.3.0
networkx>=3.2.0
